--- a/chia/full_node/block_store.py
+++ b/chia/full_node/block_store.py
@@ -58,7 +58,11 @@ class BlockStore:
         await self.db.execute("CREATE INDEX IF NOT EXISTS height on block_records(height)")
 
         await self.db.execute("CREATE INDEX IF NOT EXISTS hh on block_records(header_hash)")
-        await self.db.execute("CREATE INDEX IF NOT EXISTS peak on block_records(is_peak)")
+        
+        # this index was initially created and deployed without where condition "is_peak = 1"
+        # to replace the old index with the new one, uncomment the drop
+        # await self.db.execute("DROP INDEX IF EXISTS peak on block_records(is_peak)")
+        await self.db.execute("CREATE INDEX IF NOT EXISTS peak on block_records(is_peak) where is_peak = 1")
 
         # this index is not used by any queries, don't create it for new
         # installs, and remove it from existing installs in the future