I'd like to propose index changes.

1. Reduce size of is_peak index to 1 page by replacing it by partial index
    Described in https://github.com/Chia-Network/chia-blockchain/pull/9112#issuecomment-962682781

2. Remove group by/having clauses from this bluebox support query to improve performance
    Reduced execution times on the following systems:
    @i7-11700K, db at local SATA SSD: 
        old: 3.268s 
        new: 0.525s
    @i5-8250U, db at local mSATA SSD: 
        old: 110.811s
        new: 2.151s
    @RPi 4 8GB, db at USB3-connected NVMe SSD
        old: 222.072s
        new: 0.905s
    @kvm vm with 4 rzyen 5950x threads, qcow2 storage backed by 2x samsung  pm983 3.84 tb nvme's
        old: 35.6s
        new: 0.604s