@emlowe 2021-11-11
Thanks for the PR - I dunno what the WHERE BY HAVING was trying to do but seems your query is functional equavalent. There still seems to be some tabbing issue though
```
E     File "/home/runner/work/chia-blockchain/chia-blockchain/chia/wallet/wallet_block_store.py", line 69
E       await self.db.execute("CREATE INDEX IF NOT EXISTS peak on block_records(is_peak) where is_peak = 1")
E                                                                                                          ^
E   TabError: inconsistent use of tabs and spaces in indentation
```