@emlowe 2021-11-11
It wants you to run black on the files to get the formatting correct.
```
. ./activate
pip install ".[dev]"
black chia tests && mypy chia tests && flake8 chia tests
```
you can also just run black on the particular source file black.