@mariano54 2021-11-11
Thank you for the PR on this! Can you explain what the benchmarks are that you ran, that resulted in those large lookup times?