@neurosis69 2021-11-11
> 
> 
> It wants you to run black on the files to get the formatting correct.
> 
> ```
> . ./activate
> pip install ".[dev]"
> black chia tests && mypy chia tests && flake8 chia tests
> ```
> 
> you can also just run black on the particular source file

Found and fixed the trailing whitespaces, but got a lots of other warnings and pre-commit didn't allow me to commit.
So I just changed it in web repo - I hope it passes all these checks now.