@neurosis69 2021-11-11
> 
> 
> Thank you for the PR on this! Can you explain what the benchmarks are that you ran, that resulted in those large lookup times?

To reduce the actual output of all the heights, I put the statement as subselect and added a count.
(but I also ran it without count+subselect a couple of times, to be sure)

I used the following stmts (I think in sql trace I saw the limit of 30 to be used on bluebox request):
```
.timer on
.eqp on
select count(1) from (select height from full_blocks where is_fully_compactified=0 order by random() limit 30);
select count(1) from (SELECT height FROM full_blocks GROUP BY height HAVING sum(is_fully_compactified)=0 order by random() limit 30);
```

The old stmt uses not optimal index (full_block_height) instead of is_fully_compactified.

Runtime on my main node (i5-8250U, db at local mSATA SSD):
```
sqlite> select count(1) from (SELECT height FROM full_blocks GROUP BY height HAVING sum(is_fully_compactified)=0 order by random() limit 30);
QUERY PLAN
|--CO-ROUTINE 1
|  |--SCAN TABLE full_blocks USING INDEX full_block_height
|  `--USE TEMP B-TREE FOR ORDER BY
`--SCAN SUBQUERY 1
30
Run Time: real 117.677 user 2.633940 sys 17.054738

sqlite> select count(1) from (select height from full_blocks where is_fully_compactified=0 order by random() limit 30);
QUERY PLAN
|--CO-ROUTINE 1
|  |--SEARCH TABLE full_blocks USING INDEX is_fully_compactified (is_fully_compactified=?)
|  `--USE TEMP B-TREE FOR ORDER BY
`--SCAN SUBQUERY 1
30
Run Time: real 0.286 user 0.112955 sys 0.007778
```

Some of the other mentioned runtimes are from members of my pool helping me to compare them with my own.
Because runtimes on my local db probably doesn't compare very well, as I changed the page_size to 64k for testing purpose.