@arvidn 2021-11-12
> Found and fixed the trailing whitespaces, but got a lots of other warnings and pre-commit didn't allow me to commit.
> 
> So I just changed it in web repo - I hope it passes all these checks now.

All the tests in pre-commit are run by the CI as well, so those have to be green.