@neurosis69 2021-11-12
> 
> Regarding the indices; we need to consider the migration of existing users. With this change, as I understand it, only new users would get the smaller index.
> 
> One way to migrate would be to give the new index a different name, leave stop creating the current index (for new users), and remove the current index in some future release.

1. index named "peak":

I really wouldn't bother much, because this index is already very small and I guess we are only talking about seconds which it would take on recreating the index. But I already added the comment to drop the old index, like you suggested in my PR #9112 

Rebuild duration on my side (i5-8250U, db at local mSATA SSD + 64K page size) was as follows:

```
sqlite> select * from sqlite_master where name='peak';
index|peak|block_records|19|CREATE INDEX peak on block_records(is_peak) where is_peak = 1
sqlite> select sum(pgsize) from dbstat where name='peak';
65536
sqlite> drop index peak;
sqlite> .timer on
sqlite> CREATE INDEX peak on block_records(is_peak);
Run Time: real 2.768 user 0.289760 sys 0.478361
sqlite> select sum(pgsize) from dbstat where name='peak';
10289152
Run Time: real 0.026 user 0.020756 sys 0.005531
sqlite> drop index peak;
Run Time: real 0.042 user 0.008816 sys 0.006275
sqlite> CREATE INDEX peak on block_records(is_peak) where is_peak = 1;
Run Time: real 0.234 user 0.071092 sys 0.128647
```

2. SELECT stmt (remove group by/having) 

The other change doesn't include any actual index change.
The query optimizer automatically takes the correct index when the stmt is changed.

3. Potential next index change

There is also a potential to change the index is_fully_compactified to a partial index like:
"CREATE INDEX IF NOT EXISTS is_fully_compactified on full_blocks(is_fully_compactified where is_fully_compactified=0"

But I haven't the time right now to test/verify if it would still be needed by other chia code.
So I'll probably make a followup PR with that.