@arvidn 2021-11-12
for reference, on my machine:

```
sqlite3 ~/.chia/mainnet/db/blockchain_v1_mainnet.sqlite
sqlite> .timer on
sqlite> drop index peak;
Run Time: real 24.339 user 0.102291 sys 0.296509
sqlite> CREATE INDEX peak on block_records(is_peak);
Run Time: real 984.614 user 7.466438 sys 34.449284
```

I had to drop the page cache to benchmark creation of the smaller index.

```
sqlite> drop index peak;
Run Time: real 10.683 user 0.036494 sys 0.042373
sqlite> CREATE INDEX peak on block_records(is_peak) where is_peak = 1;
Run Time: real 1008.655 user 6.667830 sys 31.874013
```

So, I don't think it's safe to assume creating the index is fast.