@neurosis69 2021-11-12
> 
> 
> for reference, on my machine:
> 
> ```
> sqlite3 ~/.chia/mainnet/db/blockchain_v1_mainnet.sqlite
> sqlite> .timer on
> sqlite> drop index peak;
> Run Time: real 24.339 user 0.102291 sys 0.296509
> sqlite> CREATE INDEX peak on block_records(is_peak);
> Run Time: real 984.614 user 7.466438 sys 34.449284
> ```
> 
> I had to drop the page cache to benchmark creation of the smaller index.
> 
> ```
> sqlite> drop index peak;
> Run Time: real 10.683 user 0.036494 sys 0.042373
> sqlite> CREATE INDEX peak on block_records(is_peak) where is_peak = 1;
> Run Time: real 1008.655 user 6.667830 sys 31.874013
> ```
> 
> So, I don't think it's safe to assume creating the index is fast.

Very interesting. 

This would lead to some questions I have.
- which CPU your machine has and on which storage the db is located?
- was your db in sync since the beginning of mainnet? or was it synced later? (having fragmentation in mind)
- did you reorg it (vacuum) or is it untouched?
- whats the result of the following stmt on your machine?
`select sum(pgsize-unused)*100.0/sum(pgsize) from dbstat where name='block_records';`

On my machine
```
sqlite> select sum(pgsize-unused)*100.0/sum(pgsize) from dbstat where name='block_records';
97.0768199543746
Run Time: real 0.124 user 0.011323 sys 0.085690
```

I'll also ask other pool members to test, but I agree with what you said, even though I wouldn't say an index rebuild has to be done during a downtime. But it would lead to unnecessary questions.

tbh, I didn't expect that the timings you shared.