@arvidn 2021-11-12
my home directory is on an LVM volume combining 3 old spinning harddrives, circa 2006.
The machine is fast otherwise, as you can see from the timings, the vast majority of the time was spent in I/O wait.